Constrain project name during pulumi new
When picking a project name using `pulumi new`, ensure that the name
matches the regexp `^[A-Za-z0-9_.-]{1,100}$`. We'll start enforcing
this at other layers of the stack in the future (for now you could
work around these restrictions by hand editing `Pulumi.yaml`, but for
now, we'll just start restricting this to new projects created with
the wizard.